Profile UI Initial Commit
Profile UI has been redesigned. It's the final UI that needs to be built.

All of the info on profile UI is currently being displayed constantly. Need to be able to select, delete, and create new profiles.

Since profiles control the lesson unlocks, it will display the highest completed lesson at the bottom of the ui when the profile has been selected.