Finished Hiragana AI Implementation
Hiragana AI is finally implemented. The AI has an accuracy of 98% according to the training model, but in practice it's closer to 75%. Due to this, I reduced the percent required to clear a lesson to 60%, until I can increase this percentage to a better number.

The AI seems to overpredict a few Hiragana more than the others. I might go back and tweak the AI model and see where that leads.

But the Hiragana AI is implemented and ready to be used. Lessons also use the AI now, so lessons can organically be unlocked for new profiles now.