Katakana AI Implemented
The katakana AI has been fully implemented. Lessons have been slightly reworked to facilitate the use of the AI class labels. Kanji AI needs to be implemented then the project is entirely finished.

The AI performs worse than the 98.72% the notebook states, but, similar to hiragana, this is probably due to the handwritten style. I will try and tweak the layers to change this to be more generic, but it is roughly 80% accurate on average when I test it myself. It is a handwriting app, so it might be beneficial to apply these strict measures, but I would like to relax them slightly.